updated user controller

+++ app/Http/Controllers/UserController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
+use App\Http\Resources\UserResource;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
-use App\Http\Resources\UserResource;
+use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
 {
@@ -27,14 +29,20 @@ class UserController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(StoreUserRequest $request)
+    public function store(Request $request)
     {
-        $data = $request->validated();
+        $validator = $this->validation('create', $request);
+
+        if ($validator->fails()) {
+            return $this->errorResponse($validator->messages()->first(), 400);
+        }
+
+        $data = $validator->validated();
         $data['password'] = Hash::make($data['password']);
 
         $user = User::create($data);
 
-        return response()->json(new UserResource($user));
+        return $this->createdResponse(new UserResource($user));
     }
 
     /**
@@ -45,8 +53,8 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        if (! $movie = User::find($id)) {
-            return $this->core->setResponse('error', 'Movie Not Found', NULL, FALSE, 404);
+        if (! $user = User::find($id)) {
+            return $this->notFound();
         }
         return response()->json(new UserResource($user));
     }
@@ -55,28 +63,98 @@ class UserController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\User  $user
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdateUserRequest $request, User $user)
+    public function update(Request $request, $id)
     {
-        $data = $request->validated();
+        if (! $user = User::find($id)) {
+            return $this->notFound();
+        }
+
+        $validator = $this->validation('update', $request);
+
+        if ($validator->fails()) {
+            return $this->errorResponse($validator->messages()->first(), 400);
+        }
+
+        $data = $validator->validated();
         $data['password'] = Hash::make($data['password']);
 
         $user->update($data);
 
-        return response()->json(new UserResource($user));
+        return $this->successResponse(new UserResource($user));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\User  $user
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
+    public function destroy($id)
     {
+        if (! $user = User::find($id)) {
+            return $this->notFound();
+        }
+
         $user->delete();
-        return response()->json([]);
+
+        return $this->noContentResponse();
+    }
+
+    /**
+     * validation requirement
+     *
+     * @param  string $type
+     * @param  request $request
+     * @return object
+     */
+    private function validation($type = null, $request) {
+
+        switch ($type) {
+
+            case 'create':
+
+                $validator = [
+                    'name' => ['required', 'string'],
+                    'password' => ['required', 'min:6', 'max:50'],
+                    'email' => ['required', 'email', 'unique:users,email'],
+                    'role' => [
+                        'required',
+                        Rule::in(User::ROLES)
+                    ],
+                    'phone' => ['nullable', 'min:10'],
+                    'address' => ['nullable', 'string'],
+                ];
+
+                break;
+
+            case 'update':
+
+                $validator = [
+                    'name' => ['required', 'string'],
+                    'password' => ['required', 'min:6', 'max:50'],
+                    'email' => [
+                        'required',
+                        'email',
+                        Rule::unique('users', 'email')->ignore($request->get('id'))
+                    ],
+                    'role' => [
+                        'required',
+                        Rule::in(User::ROLES)
+                    ],
+                    'phone' => ['nullable', 'min:10'],
+                    'address' => ['nullable', 'string'],
+                ];
+
+                break;
+
+            default:
+
+                $validator = [];
+        }
+
+        return Validator::make($request->all(), $validator);
     }
 }
